# docs/swagger/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/v1/memberships":{"get":{"tags":["Stakeholder Membership"],"description":"Retrieve a paged collection of 5GZORRO member stakeholders according to paging and filter parameters","operationId":"getMembers","parameters":[{"name":"filterText","in":"query","required":false,"schema":{"type":"string"}},{"name":"page","in":"query","description":"Page you want to retrieve (0..N)","content":{"*/*":{"schema":{"type":"integer","default":"0"}}}},{"name":"size","in":"query","description":"Number of records per page.","content":{"*/*":{"schema":{"type":"integer","default":"20"}}}},{"name":"sort","in":"query","description":"Sorting criteria in the format: property(,asc|desc). Default sort order is ascending. Multiple sort criteria are supported.","content":{"*/*":{"schema":{"type":"array","items":{"type":"string"}}}}}],"responses":{"400":{"description":"Invalid page or filter parameters provided","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"200":{"description":"A Paged List of Member records","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Page"}}}}}},"post":{"tags":["Stakeholder Membership"],"description":"Request 5GZORRO stakeholder membership.  Request is subject to governance prior to approval.","operationId":"applyForMembership","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewMembershipRequest"}}},"required":true},"responses":{"400":{"description":"The request failed validation checks","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"200":{"description":"Request was submitted successfully","content":{"application/json":{}}}}}},"/api/v1/governance-actions":{"get":{"tags":["Governance - All Stakeholders"],"description":"Retrieve a paged collection of 5GZORRO governance proposals according to paging and filter parameters","operationId":"getProposals","parameters":[{"name":"statusFilter","in":"query","required":false,"schema":{"type":"string","enum":["PROPOSED","APPROVED","REJECTED"]}},{"name":"actionTypeFilter","in":"query","required":false,"schema":{"type":"string","enum":["ONBOARD_STAKEHOLDER","NEW_LEGAL_PROSE_TEMPLATE","ARCHIVE_LEGAL_PROSE_TEMPLATE","SLA_DISPUTE"]}},{"name":"page","in":"query","description":"Page you want to retrieve (0..N)","content":{"*/*":{"schema":{"type":"integer","default":"0"}}}},{"name":"size","in":"query","description":"Number of records per page.","content":{"*/*":{"schema":{"type":"integer","default":"20"}}}},{"name":"sort","in":"query","description":"Sorting criteria in the format: property(,asc|desc). Default sort order is ascending. Multiple sort criteria are supported.","content":{"*/*":{"schema":{"type":"array","items":{"type":"string"}}}}}],"responses":{"400":{"description":"Invalid page or filter parameters provided","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"200":{"description":"A Paged List of Proposals","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Page"}}}}}},"post":{"tags":["Governance - All Stakeholders"],"description":"Propose a governance decision. Request is subject to governance prior to approval.  The resulting identifier can be used to track the progress/status of the proposal","operationId":"proposeGovernanceDecision","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProposeGovernanceDecisionRequest"}}},"required":true},"responses":{"400":{"description":"The request failed validation checks","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"200":{"description":"Request was submitted successfully.  The proposal Id is returned in the response","content":{"application/json":{}}},"404":{"description":"The subject of the proposal (entity with provided id) could not be found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}}}}},"/api/v1/memberships/{stakeholderId}/status":{"get":{"tags":["Stakeholder Membership"],"description":"Check the status of a 5GZORRO stakeholder membership request ","operationId":"checkMembershipStatus","parameters":[{"name":"stakeholderId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Invalid identifier provided","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"404":{"description":"A member with the specified Stakeholder Id was not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"200":{"description":"Returns status information pertaining to the request","content":{"application/json":{}}}}}},"/api/v1/governance-actions/{proposalId}":{"get":{"tags":["Governance - All Stakeholders"],"description":"Retrieve a 5GZORRO governance proposal including current status information","operationId":"getGovernanceDecision","parameters":[{"name":"proposalId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Invalid identifier provided","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"404":{"description":"A proposal with the specified Id was not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"200":{"description":"Returns status information pertaining to the request","content":{"application/json":{}}}}}},"/api/v1/governance-actions/{proposalId}/vote/{accept}":{"get":{"tags":["Governance - Admin Only"],"description":"Vote on a 5GZORRO governance proposal","operationId":"voteGovernanceDecision","parameters":[{"name":"proposalId","in":"path","required":true,"schema":{"type":"string"}},{"name":"accept","in":"path","required":true,"schema":{"type":"boolean"}}],"responses":{"400":{"description":"Invalid identifier provided","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"404":{"description":"A proposal with the specified Id was not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"200":{"description":"Vote was submitted successfully","content":{"application/json":{}}}}}},"/api/v1/memberships/{stakeholderId}/revoke-membership":{"delete":{"tags":["Stakeholder Membership"],"description":"Request to revoke 5GZORRO stakeholder membership with a given stakeholder Id.  Decision to uphold the request is subject to governance if the request is not for the requesting stakeholder","operationId":"revokeMembership","parameters":[{"name":"stakeholderId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"The request failed validation checks","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}},"200":{"description":"Request was submitted successfully","content":{"application/json":{}}},"404":{"description":"A member with the specified Stakeholder Id was not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorResponse"}}}}}}}},"components":{"schemas":{"ApiErrorResponse":{"type":"object","properties":{"status":{"type":"integer","format":"int32"},"message":{"type":"string"},"errors":{"type":"object","additionalProperties":{"type":"string"}}}},"EmailNotificationDto":{"type":"object","allOf":[{"$ref":"#/components/schemas/NotificationMethodBase"},{"type":"object","properties":{"distributionList":{"type":"array","items":{"type":"string"}}}}]},"NewMembershipRequest":{"required":["notificationMethod","stakeholderClaimCertificate","stakeholderId"],"type":"object","properties":{"stakeholderId":{"type":"string"},"stakeholderClaimCertificate":{"type":"string"},"notificationMethod":{"oneOf":[{"$ref":"#/components/schemas/EmailNotificationDto"}]}}},"NotificationMethodBase":{"type":"object","properties":{"notificationType":{"type":"string","enum":["EMAIL"]}},"discriminator":{"propertyName":"notificationType"}},"ActionParamsDto":{"required":["entityIdentityId"],"type":"object","properties":{"entityIdentityId":{"type":"string"},"evidence":{"type":"string"}}},"ProposeGovernanceDecisionRequest":{"required":["actionParams","actionType"],"type":"object","properties":{"actionType":{"type":"string","enum":["ONBOARD_STAKEHOLDER","NEW_LEGAL_PROSE_TEMPLATE","ARCHIVE_LEGAL_PROSE_TEMPLATE","SLA_DISPUTE"]},"actionParams":{"$ref":"#/components/schemas/ActionParamsDto"}}},"Page":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"pageable":{"$ref":"#/components/schemas/Pageable"},"first":{"type":"boolean"},"last":{"type":"boolean"},"sort":{"$ref":"#/components/schemas/Sort"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"type":"object"}},"number":{"type":"integer","format":"int32"},"numberOfElements":{"type":"integer","format":"int32"},"empty":{"type":"boolean"}}},"Pageable":{"type":"object","properties":{"page":{"minimum":0,"type":"integer","format":"int32"},"size":{"minimum":1,"type":"integer","format":"int32"},"sort":{"type":"array","items":{"type":"string"}}}},"Sort":{"type":"object","properties":{"sorted":{"type":"boolean"},"unsorted":{"type":"boolean"},"empty":{"type":"boolean"}}}}}}